Move AppManagerWidget class into app_manager module.
To reduce the size of the app module.
